ai_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
eventlet==0.40.0
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
openai==1.79.0
packaging==25.0
pika==1.3.2
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
